Hw 04

+++ friendship-service/internal/usecase/friendship/remove_friend.go
@@ -2,6 +2,7 @@ package friendship
 
 import (
 	"context"
+	"errors"
 
 	"github.com/malytinKonstantin/go-messenger-mono/friendship-service/internal/repositories"
 )
@@ -19,6 +20,13 @@ func NewRemoveFriendUsecase(repo repositories.FriendRequestRepository) RemoveFri
 }
 
 func (uc *removeFriendUsecase) Execute(ctx context.Context, userID, friendID string) error {
+	if userID == "" {
+		return errors.New("user ID cannot be empty")
+	}
+	if friendID == "" {
+		return errors.New("friend ID cannot be empty")
+	}
+
 	err := uc.repo.DeleteFriendship(ctx, userID, friendID)
 	if err != nil {
 		return err
